+++ Week-18/weather-app/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 .env
+dist
