+++ Week-18/weather-app/src/index.ts
@@ -1,5 +1,6 @@
 require("dotenv").config();
-const app = require("./app");
+import app from  './app';
 const PORT = process.env.PORT || 4000;
 
 app.listen(PORT, () => console.log(`Server is running at port:${PORT}`));
+
