+++ Week-18/weather-app/src/routes/weather/controller.ts
@@ -1,0 +1,10 @@
+import { Request, Response } from "express";
+
+const API_URL_GEOCODE = process.env.API_URL_GEOCODE;
+const API_URL_WEATHER = process.env.API_URL_WEATHER;
+const API_KEY_NAME = process.env.API_KEY_NAME;
+const API_KEY_VALUE = process.env.API_KEY_VALUE;
+
+const getCurrentWeather = async (req: Request, res: Response) => {
+    
+};
