+++ Week-18/weather-app/src/routes/weather/index.ts
@@ -1,0 +1,7 @@
+import express from "express";
+
+const router = express.Router();
+
+router.route("/").get();
+
+export default router;
